nep245 checked event gated with near-contract
