+++ crates/near/nep245/src/lib.rs
@@ -1,8 +1,10 @@
+#[cfg(feature = "near-contract")]
 mod checked;
 #[cfg(feature = "near-contract")]
 mod core;
 #[cfg(feature = "near-contract")]
 pub mod enumeration;
+mod error;
 mod events;
 #[cfg(feature = "near-contract")]
 pub mod receiver;
@@ -10,11 +12,7 @@ pub mod receiver;
 pub mod resolver;
 mod token;
 
-pub use self::{
-    checked::{CheckedMtEvent, ErrorLogTooLong},
-    events::*,
-    token::*,
-};
+pub use self::{error::ErrorLogTooLong, events::*, token::*};
 
 #[cfg(feature = "near-contract")]
 pub use self::core::*;
